Unformat CarouselViewRenderer, recombine files; no code changes
/*privat*/ -> #region

Unformat CarouselView.cs

More unformat work

Remove offending /*privates*/
